스토어의 상품 추가 내역을 저장하는 테스트코드 작성

+++ src/main/java/com/gigacoffeebackend/history/domain/ProductHistoryCollection.java
@@ -12,8 +12,8 @@ import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 import org.springframework.data.mongodb.core.mapping.MongoId;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Document("product_histories")
 @Getter
@@ -37,7 +37,7 @@ public class ProductHistoryCollection {
     private String productName;
 
     @Field("product_price")
-    private int productPrice;
+    private BigDecimal productPrice;
 
     @Field
     @CreatedDate
